refactor: move step 2 of submit to own file
