+++ src/actions/submit/submit.ts
@@ -9,7 +9,6 @@ import { ExitFailedError, PreconditionsFailedError } from '../../lib/errors';
 import { cliAuthPrecondition } from '../../lib/preconditions';
 import { getSurvey, showSurvey } from '../../lib/telemetry/survey/survey';
 import {
-  detectUnsubmittedChanges,
   gpExecSync,
   logError,
   logInfo,
@@ -21,35 +20,18 @@ import { assertUnreachable } from '../../lib/utils/assert_unreachable';
 import { Unpacked } from '../../lib/utils/ts_helpers';
 import { Branch } from '../../wrapper-classes/branch';
 import { TScope } from '../scope';
-import { getPRBody } from './pr_body';
-import { getPRDraftStatus } from './pr_draft';
-import { getPRTitle } from './pr_title';
-import { getReviewers } from './reviewers';
+import { getPRInfoForBranches } from './prepare_branches';
 import { getValidBranchesToSubmit } from './validate_branches';
 
 export type TSubmitScope = TScope | 'BRANCH';
-
-type TPRSubmissionAction = { branch: Branch; parent: Branch } & (
-  | { update: false }
-  | {
-      update: true;
-      prNumber: number;
-    }
-);
 
 type TPRSubmissionInfo = t.UnwrapSchemaMap<
   typeof graphiteCLIRoutes.submitPullRequests.params
 >['prs'];
 
-type TSubmittedPRRequest = Unpacked<
+export type TSubmittedPRRequest = Unpacked<
   t.UnwrapSchemaMap<typeof graphiteCLIRoutes.submitPullRequests.params>['prs']
 >;
-
-type TSubmittedPRRequestWithBranch = TSubmittedPRRequest & {
-  branch: Branch;
-};
-
-type TPRSubmissionInfoWithBranch = TSubmittedPRRequestWithBranch[];
 
 type TSubmittedPRResponse = Unpacked<
   t.UnwrapSchemaMap<typeof graphiteCLIRoutes.submitPullRequests.response>['prs']
@@ -103,12 +85,6 @@ export async function submitAction(
   }
 
   // Step 2: Prepare
-  logInfo(
-    chalk.blueBright(
-      '🥞 [Step 2] Preparing to submit PRs for the following branches...'
-    )
-  );
-
   const submissionInfoWithBranches = await getPRInfoForBranches(
     {
       branches: branchesToSubmit,
@@ -183,106 +159,6 @@ async function submitPullRequests(
 
   saveBranchPRInfo(prInfo, context);
   printSubmittedPRInfo(prInfo);
-}
-
-/**
- * For now, we only allow users to update the following PR properties which
- * necessitate a PR update:
- * - the PR base
- * - the PR's code contents
- *
- * Notably, we do not yet allow users to update the PR title, body, etc.
- *
- * Therefore, we should only update the PR iff either of these properties
- * differ from our stored data on the previous PR submission.
- */
-async function getPRInfoForBranches(
-  args: {
-    branches: Branch[];
-    editPRFieldsInline: boolean;
-    draftToggle: boolean | undefined;
-    updateOnly: boolean;
-    dryRun: boolean;
-    reviewers: boolean;
-  },
-  context: TContext
-): Promise<TPRSubmissionInfoWithBranch> {
-  return await Promise.all(
-    args.branches
-      .map((branch) => {
-        // The branch here should always have a parent - above, the branches we've
-        // gathered should exclude trunk which ensures that every branch we're submitting
-        // a PR for has a valid parent.
-        const parent = branch.getParentFromMeta(context);
-        if (parent === undefined) {
-          throw new PreconditionsFailedError(
-            `Could not find parent for branch ${branch.name} to submit PR against. Please checkout ${branch.name} and run \`gt upstack onto <parent_branch>\` to set its parent.`
-          );
-        }
-        const prNumber = branch.getPRInfo()?.number;
-
-        const status =
-          prNumber === undefined
-            ? args.updateOnly
-              ? 'NOOP'
-              : 'CREATE'
-            : branch.isBaseSameAsRemotePr(context)
-            ? 'RESTACK'
-            : detectUnsubmittedChanges(branch)
-            ? 'CHANGE'
-            : args.draftToggle === undefined
-            ? 'NOOP'
-            : args.draftToggle
-            ? 'DRAFT'
-            : 'PUBLISH';
-
-        logInfo(
-          `▸ ${chalk.cyan(branch.name)} (${
-            {
-              CREATE: 'Create',
-              CHANGE: 'Update - code changes/rebase',
-              DRAFT: 'Convert to draft - set draft status',
-              NOOP: 'No-op',
-              PUBLISH: 'Ready for review - set draft status',
-              RESTACK: 'Update - restacked',
-            }[status]
-          })`
-        );
-        return args.dryRun || status === 'NOOP'
-          ? undefined
-          : {
-              branch,
-              parent,
-              ...(status !== 'CREATE'
-                ? { update: true, prNumber }
-                : { update: false }),
-            };
-      })
-      .filter((action): action is TPRSubmissionAction => action !== undefined)
-      .map(async (action) => {
-        return action.update
-          ? {
-              action: 'update' as const,
-              prNumber: action.prNumber,
-              draft: args.draftToggle,
-              head: action.branch.name,
-              headSha: action.branch.getCurrentRef(),
-              base: action.parent.name,
-              baseSha: action.branch.getParentBranchSha(),
-              branch: action.branch,
-            }
-          : await getPRCreationInfo(
-              {
-                branch: action.branch,
-                parentBranchName: action.parent.name,
-                editPRFieldsInline: args.editPRFieldsInline,
-                draftToggle: args.draftToggle,
-                reviewers: args.reviewers,
-              },
-              context
-            );
-      })
-  );
 }
 
 function pushBranchesToRemote(branches: Branch[], context: TContext): Branch[] {
@@ -405,61 +281,6 @@ async function pushMetaStacks(branchesPushedToRemote: Branch[]): Promise<void> {
       }
     );
   });
-}
-
-async function getPRCreationInfo(
-  args: {
-    branch: Branch;
-    parentBranchName: string;
-    editPRFieldsInline: boolean;
-    draftToggle: boolean | undefined;
-    reviewers: boolean;
-  },
-  context: TContext
-): Promise<TSubmittedPRRequestWithBranch> {
-  logInfo(
-    `Enter info for new pull request for ${chalk.yellow(args.branch.name)} ▸ ${
-      args.parentBranchName
-    }:`
-  );
-
-  const submitInfo = {
-    title: await getPRTitle(
-      {
-        branch: args.branch,
-        editPRFieldsInline: args.editPRFieldsInline,
-      },
-      context
-    ),
-    body: await getPRBody(
-      {
-        branch: args.branch,
-        editPRFieldsInline: args.editPRFieldsInline,
-      },
-      context
-    ),
-    reviewers: await getReviewers({
-      fetchReviewers: args.reviewers,
-    }),
-  };
-  args.branch.upsertPriorSubmitInfo(submitInfo);
-
-  const createAsDraft = args.draftToggle ?? (await getPRDraftStatus());
-
-  // Log newline at the end to create some visual separation to the next
-  // interactive PR section or status output.
-  logNewline();
-
-  return {
-    ...submitInfo,
-    action: 'create',
-    draft: createAsDraft,
-    head: args.branch.name,
-    headSha: args.branch.getCurrentRef(),
-    base: args.parentBranchName,
-    baseSha: args.branch.getParentBranchSha(),
-    branch: args.branch,
-  };
 }
 
 function printSubmittedPRInfo(prs: TSubmittedPR[]): void {
